python-ai/requirements.txt
flask
flask-cors
opencv-python
pillow
numpy
python-dotenv==1.0.0
